restructure/clean up `extract` subcommand.
